refactor(supervisor): split into domain modules under supervisor/
Pure-move split of the 3,198-line supervisor.rs along its domain seams:

- mod.rs         Supervisor state + status state machine + turn-end hub
- events.rs      event payload structs + emit helpers
- lifecycle.rs   spawn/resume/switch/respawn/stop + process wiring + watchdogs
- disposition.rs archive/restore/discard + worktree snapshot/teardown
- messaging.rs   message delivery matrix, queue flush/drain paths
- shell.rs       per-agent interactive shell PTY
- run.rs         Run-panel phase planning + chaining
- session_sync.rs turn-end transcript ingest + PR state
- rpc_watch.rs   RPC mailbox watcher (loop body extracted as handle_rpc_event)

All code and comments moved verbatim; only use/visibility/module
declarations changed. lib.rs and commands.rs are untouched. Tests move
with their subjects; messaging reuses mod.rs's cfg(test) helpers.
